requirements.txt
cssmin==0.2.0
Flask==0.10.1
Flask-Assets==0.11
gunicorn==19.4.5
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
libsass==0.10.1
MarkupSafe==0.23
oauth2==1.9.0.post1
six==1.10.0
webassets==0.11.1
Werkzeug==0.11.4
yelp==1.0.1
